+++ processor/src/test/fixtures/input/com/example/inheritance/complex_interface/InnerInterface.java
@@ -2,6 +2,7 @@ package com.example.inheritance.complex_interface;
 
 import react4j.annotations.Input;
 
+@SuppressWarnings( "React4j:MethodBasedImmutableInput" )
 interface InnerInterface
   extends RootInterface
 {