+++ processor/src/test/fixtures/input/com/example/prop/ImmutableDisposableArezComponentMaybeTrackingProp.java
@@ -8,6 +8,7 @@ import react4j.annotations.Input;
 import react4j.annotations.Render;
 import react4j.annotations.View;
 
+@SuppressWarnings( "React4j:MethodBasedImmutableInput" )
 @View( type = View.Type.MAYBE_TRACKING )
 abstract class ImmutableDisposableArezComponentMaybeTrackingProp
 {