+++ processor/src/test/fixtures/input/com/example/prop/ImmutablePropTypeKeyedInterface.java
@@ -7,6 +7,7 @@ import react4j.annotations.Input;
 import react4j.annotations.Render;
 import react4j.annotations.View;
 
+@SuppressWarnings( "React4j:MethodBasedImmutableInput" )
 @View
 abstract class ImmutablePropTypeKeyedInterface
 {