+++ processor/src/test/fixtures/input/com/example/prop/SuppressedReact4jMethodImmutableInputWarningModel.java
@@ -4,19 +4,16 @@ import javax.annotation.Nullable;
 import react4j.ReactNode;
 import react4j.annotations.Input;
 import react4j.annotations.Render;
+import react4j.annotations.SuppressReact4jWarnings;
 import react4j.annotations.View;
 
-@SuppressWarnings( "React4j:MethodBasedImmutableInput" )
+@SuppressReact4jWarnings( "React4j:MethodBasedImmutableInput" )
 @View
-abstract class ImmutableMixedInputModel
+abstract class SuppressedReact4jMethodImmutableInputWarningModel
 {
-  ImmutableMixedInputModel( @Nullable @Input( immutable = true ) final String myImmutableProp )
-  {
-  }
-
   @Nullable
-  @Input
-  abstract String getMyMutableProp();
+  @Input( immutable = true )
+  abstract String getMyProp();
 
   @Nullable
   @Render